backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
opencv-python-headless==4.10.0.84
mediapipe
numpy
google-generativeai==0.7.2
sqlalchemy==2.0.36
python-dotenv==1.0.1
httpx==0.27.0
scipy